Backend/requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
django-debug-toolbar
django-cors-headers
django-filter
django-taggit
drf-spectacular
django-environ
markdown
Pillow
